Replace 'config.Datadog' global by a function to retrieve the config
The globals 'pkg/config.Datadog' and 'pkg/config/setup.Datadog' have
been replace by a function 'Datadog()' to access the global config.

This is the first step toward preventing a race condition accessing the
config before initialization was possible. This also pave the way toward
a complete migration of pkg/config to component.

Finally, this removes the last blocker for the creation of a dedicated
package for pkg/config/mock.go.