Change testing of configprovider
Currently, the config returned by the custom config provider Get is tested against a build out struct. This will become unmaintainable as we add more test cases. This PR changes this so we test the output of the custom Get against the output of Get using the "upstream" configprovider. Currently, we are not modifying the config, so both Gets point to the same URI. Once we add support for enhanced config, we will start pointing the "upstream" get to "result.yaml" files.