Replace github.com/mholt/archiver/v3 third party library with custom package pkg/util/archive
The libray includes the vulnerability CVE-2024-0406. Technically speaking we are not affected by this in production as we do not uncompress tar files. Still, it is a good opportunity to evaluate if we need a third party library for zip opertaions.

We decided to remove the third party dependency by vendoring some of its code.

That has the benefit that we are no longer affected by the CVE-2024-0406 and that we not longer depend on a third party dependency.