[USMON-358] postgres monitoring (#25416)
* usm: Update agent-payload version

Updating to a version that supports DB aggregations from USM

* usm: postgres: Added configuration to enable postgres monitoring

* usm: postgres: Added configuration to control max stats buffered

Every protocol has its own configuration to control the number of stats in memory
to store until the next connections check

* usm: postgres: Added eBPF decoding code

The commit contains the postgres decoding plaintext implementation.

* usm: postgres: Added protocol

The commit adds the postgres protocol struct to USM.
The commit does not contain actual user-mode processing of postgres events. That part will be
added in a dedicated commit

* usm: postgres: Add test build flag, to test utils files

* usm: postgres: Add statskeeper and processing of events

* usm: postgres: Expose postgres requests to connections check

* usm: postgres: Added debugging endpoint

* Fixed CR comments

* usm: postgres: Added postgres encoder

* usm: postgres: Added UTs

* usm: postgres: Handle verifier error on old kernels

* usm: postgres: Fixed wrong latnecy for multiple records

If we have a single record, then we should use firstLatencySample, but if we have
many records, we should use the latency field

* usm: postgres: Update protobuf structs

* usm: postgres: tests: Address CR comments