usm: kafka: Add TLS support (#25083)
* usm: kafka: Enable TLS on server

* usm: kafka: Introduce pktbuf abstraction

In order to avoid copy-pasting the code to support TLS, add an
abstraction called pktbuf which wraps the helpers necessary to access
the packet and the metadata depending on whether it is backed by an SKB
or a user space TLS buffer in a unified interface.

This abstraction is intended to be zero-cost.  We rely on the compiler
to optimize away the unused code paths and structure passing.  The
kafka_response_parser program is fully identical before and after this
change.  The kafka_filter program sees a growth of 7 bytes (not 100%
sure why).

* usm: kafka: Handle split response metadata

TLS libraries may choose to read the message size and/or the correlation
ID from the fetch response separately, and these will show up as
separate "packets" in our hooks.  Currently, we assume that the metadata
and the header will be in one packet, so rework this check.

* usm: kafka: Add TLS support

* usm: testutil: Add control messages to unix proxy

Precise control over the sizes of the read/writes performed is needed to
test packet split handling.  Add control messages to the unix proxy
which send the size of the read/write to be performed.  The default mode
of the proxy is still to be transparent; the new control messages will
initially only be used from the upcoming Kafka tests.

These control messages should allow avoiding having to add a sleep
between write calls (as present in the HTTP/2 tests).

* usm: kafka: Run tests with and without TLS

Add support for running the tests with TLS too.  In order to do this,
convert the tests to use the unix proxy since the unit test binary
itself is built without symbols and cannot be hooked.  For simplicity,
the unix proxy is used both in the cases with and without TLS.

* usm: Move READ_INTO_BUFFER_USER to correct header

Move it to the read_into_buffer.h header so that we don't need to
include http/buffer.h from Kafka code (and thus avoid getting errors due
to unused functions in some builds).

* usm: kafka: Add __maybe_unused to a few functions

* usm: tls: Use correct PROG_MAX value

* usm: Add comment to tls_dispatcher_classification_progs

* usm: big-endian: Fix indentation

* usm: kafka: test: Move server address to common struct

* usm: kafka: test: Clarify TLS condition

* usm: kafka: test: Add comment for CannedClientServer

* usm: tls: Fix type of TLSDispatcherKafkaProg

* usm: read_into_buffer: Remove unnecessary whitespace/continuation

* ebpf: Decouple pktbuf from kafka

* usm: pktbuf: Fix formatting of macros

* usm: pktbuf: Use skb->len if no skb_info

* usm: pktbuf: Use telemetry version of read_user()