[ASCII-1615] Replace `github.com/mholt/archiver/v3` third party library with custom package `pkg/util/archive`  (#25498)
* Replace github.com/mholt/archiver/v3 third party library with custom package pkg/util/archive

The libray includes the vulnerability CVE-2024-0406. Technically speaking we are not affected by this in production as we do not uncompress tar files. Still, it is a good opportunity to evaluate if we need a third party library for zip opertaions.

We decided to remove the third party dependency by vendoring some of its code.

That has the benefit that we are no longer affected by the CVE-2024-0406 and that we not longer depend on a third party dependency.

* Add tests for archive.Zip and archive.UnZip

* Add copyright lines

* PR feedback

* ensure file is closed when the test finish

* Ensure we write the Zip header file for folders

* use SecureJoin to extract file from archive