USM: ssl: Ignore ELF files for other architectures (#25505)
* USM: ssl: Ignore ELF files for other architectures

Ignore ELF files which are not for the architecture we're running on.
Without this, we could end up installing a uprobe on, for example, an
arm64 binary on an amd64 machine, thus corrupting the arm64 instruction
and leading to segmentation faults.

* Use blockList in test

* Use OpenFromAnotherProcess instead of libmmap

* Remove unused libmmap

* Simplify testArch

* Skip test on unsupported platforms

* Format fakessl.c with clang-format

* Add shebang link to script

* Move libs to testdata

So that kmt will copy it to the VMs.