system-probe: Add pkg/process/monitor to tests (#25463)
* system-probe: Add pkg/process/monitor to tests

The process monitor used by USM lives under pkg/process/monitor and has
a test suite; run it as part of the system-probe tests to prevent it
from breaking.

* system-probe: Fix formatting with ruff

* system-probe: Ignore packages without dirs

If the package does not contain any dirs for the current set of tags,
ignore it instead of reporting empty paths which cause crashes later.
This is needed now since pkg/process/monitor is not available for
Windows.

* Use explicit check

Co-authored-by: Guillermo Julián <gjulianm@users.noreply.github.com>

---------

Co-authored-by: Guillermo Julián <gjulianm@users.noreply.github.com>