pkg/config/model: Update the Viper config to copy proxy settings.
The Viper config previously did not copy the proxies in the CopyConfig
function, and in addition, the sync.Once could not be reset and would
prevent subsequent correct usage of the proxy field.

This PR copies the proxies field and also removes the Once, instead
utilizing the config mutex to ensure serial access.