usm: tests: Split classification tests to linux only and cross OS
Most of the tests are linux only and there is no matching classification utility
in the windows version. Today the tests are shared and run for both OS but skipped
on windows. That complicates writing tests and does not allow to easily share
utilities implemented in a linux only file as it will break the windows CI.

The PR moves the linux only classification tests into a linux file, and split
the ancestor test into CrossOS and Linux composers