pkg/trace/agent: remove trace/stats writer channels (#25602)
The trace writer and stats writer both currently ingest their data via a
channel which they then process with their own goroutines. These
additional routines allow for trace data to exist longer and make it
harder to detect when to apply backpressure, to avoid OOMs.

The coordination / delay related to using a channel + go routine to
process this data separately also adds a small amount of CPU / Memory
overhead (we will measure with more benchmarks, initial macro benchmark
results have been good).

An additional benefit is some of the unit tests here get a lot simpler as
they no longer need to worry about async processing of written traces,
and can make use of a mock trace writer (or noop trace writer if they
don't need to assert on the output).

This should be slightly more performant, but it's possible there are some
load patterns that may be worse - we will measure a variety of traffic
patterns / max throughput to be sure here.

These changes will require updates from users of these components, namely
open telemetry contrib (however! this may also be a benefit as moving to
an interface for trace and stats writing makes it easier for OTEL to
provide their own implementation or use a NOOP implementation if desired)