usm: kafka: Support Fetch v12
Add support for v12 of the Fetch API.  The main change is the use of
varints and addition of tagged values.  While that change itself is
fairly straightforward, the need to parse varints caused the programs to
exceed the instruction size supported by older kernels (4096
instructions), so the programs had to be split into separate programs
for pre-v12 and v12, as well as to do the parsing in two steps in
separate programs: a first step to parse the locations of the record
batches arrays, and a second step to parse the list of found locations.

The split between pre-v12 and v12 is simple and handled by informing the
compiler about the bounds of the versions, and then depending on it to
eliminate the unused branches.  The split of the parsing into two steps
requires code to handle the transition between the two steps correctly.

Also mainly due to code size restrictions on older kernels (even after
the above optimizations), the following limitations are currently
present:

 - A maximum of four bytes of varints are supported
 - The tagged values fields are always assumed to be empty