pkg/trace/writer: reduce sender count and improve retry
This change reduces the default number of concurrent senders from 120 to
5. Having 120 simultaneous senders means that 120 payloads can be
in-flight at a given time, which is far too high considering the amount
of memory it can require (120 * ~10MiB = ~1.2GiB)

An I/O-bound operation like sending a payload does not benefit from high
concurrency, so reducing this number should not affect throughput.

This change also improves the retry mechanism by having a per-payload
backoff, and retrying payloads as they fail, rather than attempting to
requeue them, which can result in dropping retryable payloads when the
queue is full.