[windows][CWS] Modify channel behavior.
First of several performance fixes.

Modifies the channel used to submit events to the CWS processor.
Was a blocking, unbuffered channel.  This is now runtime configurable.
The configuration is not intended to be changed by customers, but to make it easier
to tune with a/b testing internally.

Overall changes:
- defaults to non-blocking channel.  This prevents events from backing up
  into ETW.  Also allows clearer notion of what events were dropped.  When
  they back up into ETW, all we know is a count of dropped events;

- sets a default channel size.  This allows a small amount of buffering to
  smooth out slight bursts in ETW notifications.  The default right now
  is fairly small to prevent large memory usage.  intended to be tuned over time.

- allows ability to configure etw max buffers.  This is also for tuning,
 and our default (0) is to use the system default.