[windows][CWS] Update the statistics that are collected. (#25919)
* [windows][CWS] consolidate windows probe initialization.

Refactor initialization of windows probe.

Solves problem that each time we add new structure member, needed to remember
to copy/paste the initialization to the tests, too.

* fix unneeded import in test

* fix initialization of tests

* [windows][CWS] Update the statistics that are collected.

Add a general map for collecting which etw messages we receive and
which ones we actually process.

use a more general map so we don't have to update the stat every time
we add/remove one.

* ....