[windows][cws] add approvers (#25924)
* [windows][CWS] consolidate windows probe initialization.

Refactor initialization of windows probe.

Solves problem that each time we add new structure member, needed to remember
to copy/paste the initialization to the tests, too.

* fix unneeded import in test

* fix initialization of tests

* [windows][CWS] Update the statistics that are collected.

Add a general map for collecting which etw messages we receive and
which ones we actually process.

use a more general map so we don't have to update the stat every time
we add/remove one.

* [windows][CWS] Modify channel behavior.

First of several performance fixes.

Modifies the channel used to submit events to the CWS processor.
Was a blocking, unbuffered channel.  This is now runtime configurable.
The configuration is not intended to be changed by customers, but to make it easier
to tune with a/b testing internally.

Overall changes:
- defaults to non-blocking channel.  This prevents events from backing up
  into ETW.  Also allows clearer notion of what events were dropped.  When
  they back up into ETW, all we know is a count of dropped events;

- sets a default channel size.  This allows a small amount of buffering to
  smooth out slight bursts in ETW notifications.  The default right now
  is fairly small to prevent large memory usage.  intended to be tuned over time.

- allows ability to configure etw max buffers.  This is also for tuning,
 and our default (0) is to use the system default.

* [windows][cws] restrict ETW message flow

Another performance enhancement, this PR builds on #25875 and
limits the messages that ETW will send us to the once we're specifically
looking to handle. Significantly reduces load on the system-probe

* [windows][cws] Re-add approvers for windows

This is a resubmit of a previous PR to implement basic approvers for windows.

* fix space

* lint