usm: kafka: test: Make checks more robust (#25969)
* usm: kafka: test: Simplify Eventually checks

Use EventuallyWithT() which allows to assert different conditions
separately and thus prints better errors messages (such as the compared
values in case the assertion fails).  Also, get rid of the PrintableInt
usage since it does not add value here.

* usm: kafka: test: Combine stats wait and validation

The separation of waiting for the stats and verifying them in two
different steps is racy for the case when there are multiple events
involved (eg. multiple requests/responses) since if the events get
flushed (for example, due to the flush hook in the netif_receive_skb
tracepoint), then the first condition will be satisfied but not the
second one.