usm: kafka: Support fetch v12 (#25812)
* usm: kafka: Use unsigned loop counter

This allows some compiler optimzations and avoids a "looks like the BPF
stack limit of 512 bytes is exceeded" error when more code is added to
the classification functions.

* usm: kafka: test: Make unused fields non-zero

Set some unparsed fields in the packets to non-zero values for easier
debugging when packet parsing goes wrong and reads from wrong offsets.

* usm: kafka: test: Pass FetchRequest to buildMessages

To allow the request to be customized per test.

* usm: kafka: test: Pass in API version

Pass in the API version as a parameter to the raw tests, so that
different API versions can be tested.

* usm: kafka: test: Test different versions in raw test

Test some versions which have specific conditions.

* usm: kafka: test: Allow topic to be customized

* usm: kafka: test: Add test for large topic name

* usm: kafka: test: Add test for many topics

* usm: kafka: test: Add raw test for many partitions

* usm: kafka: test: Handle flexible header

* usm: kafka: Pass in version number to parsing tests

* usm: kafka: Support Fetch v12

Add support for v12 of the Fetch API.  The main change is the use of
varints and addition of tagged values.  While that change itself is
fairly straightforward, the need to parse varints caused the programs to
exceed the instruction size supported by older kernels (4096
instructions), so the programs had to be split into separate programs
for pre-v12 and v12, as well as to do the parsing in two steps in
separate programs: a first step to parse the locations of the record
batches arrays, and a second step to parse the list of found locations.

The split between pre-v12 and v12 is simple and handled by informing the
compiler about the bounds of the versions, and then depending on it to
eliminate the unused branches.  The split of the parsing into two steps
requires code to handle the transition between the two steps correctly.

Also mainly due to code size restrictions on older kernels (even after
the above optimizations), the following limitations are currently
present:

 - A maximum of four bytes of varints are supported
 - The tagged values fields are always assumed to be empty

* usm: kafka: test: Check for errors from LookupMaxKeyVersion

* usm: kafka: Remove leftover comment

* usm: kafka: remove unused cgo type

* usm: kafka: Remove passing of level

* usm: kafka: Add comment to parse_varint_u16()

* usm: kafka: Fix u16 varint check

The second byte's MSB should only be checked if the first byte has the
MSB set, since otherwise the second byte is unrelated to the varint.

* usm: kafka: Rename topic_name_size_tmp to topic_name_size_raw

* usm: kafka: Handle splits in header

We were previously not handling this mainly due to code size
restrictions, but with the program split this can be supported.

* usm: kafka: Make read_with_remainder generic

Remove the copy/pasted code and make read_with_remainder() generic with
the use of a function pointer.  The function pointer will be folded at
compile time by the compiler.  The instruction count is unchanged before
and after this patch.

* usm: kafka: Remove unused nonrestartable functions

* usm: kafka: Remove _restartable suffix

Now that we only have restartable readers, remove the redundant suffix.

* usm: kafka: Validate zero tagged fields when possible

* usm: kafka: Remove __read_varint indirection

* usm: kafka: Allow varint bytes to be customized per location

To reduce code size by limiting iterations wherever it makes sense.

* usm: kafka: Customize varint bytes per location

* usm: kafka: Clarify N-1 operation in varints

* usm: kafka: Rename tmp2 variable

* usm: kafka: Add comment about flexible

* usm: kafka: Skip varint number of topics

* usm: kafka: Check zero tagged fields in request

* usm: kafka: Use 64 bits for varint results

* usm: kafka: Add comment about uninitialized variable

* usm: kafka: test: Fix teardown

The telemetry test assigns to the client in the context multiple times,
so some of the clients are never torn down, leading to issues with tests
run later (such as for example the still-running clients attempting to
connect to the raw server while the raw tests are executed afterward).