[windows][cws] Fix cleaning up discarders (#26112)
This change fixes the issue where we're never removing the file handle
from the discarded handles cache.  So, once a file handle is discarded,
it stays discarded even if that handle is closed, and then is reused on
subsequent file open.

Also, since we could miss the close operation, if we see a new handle,
remove it from the discarded cache.