pkg/trace/writer: reduce sender count and improve retry (#25800)
* pkg/trace/agent: remove trace/stats writer channels (#25602)

The trace writer and stats writer both currently ingest their data via a
channel which they then process with their own goroutines. These
additional routines allow for trace data to exist longer and make it
harder to detect when to apply backpressure, to avoid OOMs.

The coordination / delay related to using a channel + go routine to
process this data separately also adds a small amount of CPU / Memory
overhead (we will measure with more benchmarks, initial macro benchmark
results have been good).

An additional benefit is some of the unit tests here get a lot simpler as
they no longer need to worry about async processing of written traces,
and can make use of a mock trace writer (or noop trace writer if they
don't need to assert on the output).

This should be slightly more performant, but it's possible there are some
load patterns that may be worse - we will measure a variety of traffic
patterns / max throughput to be sure here.

These changes will require updates from users of these components, namely
open telemetry contrib (however! this may also be a benefit as moving to
an interface for trace and stats writing makes it easier for OTEL to
provide their own implementation or use a NOOP implementation if desired)

* pkg/trace/writer: reduce sender count and improve retry

This change reduces the default number of concurrent senders from 120 to
5. Having 120 simultaneous senders means that 120 payloads can be
in-flight at a given time, which is far too high considering the amount
of memory it can require (120 * ~10MiB = ~1.2GiB)

An I/O-bound operation like sending a payload does not benefit from high
concurrency, so reducing this number should not affect throughput.

This change also improves the retry mechanism by having a per-payload
backoff, and retrying payloads as they fail, rather than attempting to
requeue them, which can result in dropping retryable payloads when the
queue is full.

* cleanup

* Update pkg/trace/writer/sender.go

---------

Co-authored-by: Andrew Glaude <andrew.glaude@datadoghq.com>