usm: kafka: test: Reduce test size
The TestKafkaFetchRaw/with_TLS/api*/split/many_partitions fails
sometimes in CI with an error of the type with an error message similar
to "Expected 100000 fetch requests but got 99900", where the "but got"
number varies.  The problem has not been reproduced locally yet, but
reduce the counts in an attempt to limit the flakiness.