usm: Remove TLS-specific program enums
The TLS-specific program enums are almost identical to the non-TLS ones,
the only difference being the TERMINATION values and some naming
inconsistency (eg. HTTP2_FILTER vs HTTP2_FRAME_FILTER).  Eliminate
these and just use the non-TLS enums for the TLS program array too.  The
only cost is the increase in four entries (corresponding to the
TERMINATION ones) in the normal program array.

This will make it easier to add new buffer types without having to
duplicate this enum again.