[APM] Add a default UDS listener for traces
This PR adds a default UDS listener on `/var/run/datadog/apm.socket`.
This is used in the k8s case and SSI on host/docker rely on this socket
path.

If the user-configured socket differs from this default one, 2 different
listeners will be created.

If "/var/run/datadog" is not created by systemd (on very old systems),
there is no need to create this socket.