(installer): pin default versions of APM libraries
In the effort to avoid breaking changes for customers we are proposing a
default pin of the library versions. In addition to this change we will
be updating the default install command in the UI and public
documentation to include any new versions of libraries.

By pinning to the major tag: libraries can still release bug fixes and
add new features to customers; and customer will not receive breaking
changes (like dropped support of runtimes/integrations).