fix(serverless/appsec): language runtime proxy not started if no API Key is available
When no Datadog API key is available (either because it is not
configured, or because it cannot be accessed/decrypted), the serverless
agent runs in "no-op" mode. However in this case, the language runtime
proxy is never started, even though the function's configuration may
require it. This is the case when Serverless AppSec is enabled.

This changes the noop behavior to also start the AppSec language runtime
API proxy to satisfy the function configureation; if AppSec would have
started in normal conditions (with an API key).