usm: tests: Reuse event stream and process monitoring over a test suite
We setup event stream monitoring (if needed) and process monitor only once per a suite run
In the leaf tests we register and de-register dedicated callbacks to verify the tests are
working. Thus, we don't have to create over and over again the event monitoring for every
leaf test.

The motivation is the setup/teardown time for the event monitoring, which takes 10s.
Thus, running event monitoring tests are very expensive