usm: Only create event consumer if used
Currently, the process monitor event stream consumer is registered
if EnableUSMEventStream is set, but there are other options which
determine if the process monitor really is used.  If the process monitor
is not used and the event stream consumer is registered, there is a
segfault when events are received since the process monitor has not been
initialized:

 panic: runtime error: invalid memory address or nil pointer dereference
 [signal SIGSEGV: segmentation violation code=0x1 addr=0x0 pc=0x2eaba1c]
 goroutine 2168 [running]:
 pkg/network/protocols/telemetry.(*Counter).Add(...)
 pkg/process/monitor.(*EventConsumer).HandleEvent()
 pkg/security/probe.(*EventConsumer).Start.func1()
 created by pkg/security/probe.(*EventConsumer).Start in goroutine 1

To fix this, ensure that the conditions for registering the event stream
consumer and initializing the process monitor are the same.