usm: http2: Added flush from the uprobes
Flushing happens in a kprobe 'tracepoint__net__netif_receive_skb', the kprobe has high miss count
which means it might run very late (or none at all) on the same cpu the uprobe was executed on.
That means, we can have entries stuck in the batches map (which is per-cpu), that is never
being populated. That can explain the accuracy misses in the CI