usm: kafka: test: Run split test in groups
The split tests tests a lot of transactions and this causes the event
flushing to sometimes fail with tests generating multiple events per
transaction like the "with different error codes" test because there is
no room in the perf buffer.  Split the test into groups to reduce the
risk of this happening.