usm: tests: Added a manual attachment to a process as a fallback
If specified, and the test does not intended to test the attachement process, but the capturing of TLS traffic,
then we wait to see if the monitor hooks the given PID, if it didn't and the fallback is enabled, then we manually
attach to the process.

This is intended to handle flakiness in tests (that are not related to the attachement feature) like http2 monitoring.