stop manager before protocols
This is to prevent a send on a closed channel in the event consumers. If the manager is not stopped, it will continue to consume from the ebpf perf/ring buffers and call callbacks. Those callbacks will try to send on a closed channel due to a race on the `closed` flag.

Technically this is a data race on the `closed` flag, but there isn't a good way to prevent it without having some performance impact.