usm: classificiation: tls: Convery network byte order to host byte order
The two fields are sent as network order but we need to convert them to host order, for true
comparison to the consts we support.

Since we call bpf_ntohs on the struct, using a pointer cast to the original buffer, can change the original
buffer, and impact other classifiers. Instead, we create a local copy of the struct so we can safely call
bpf_ntohs