[apm] aggregate _dd.base_service tag for internal service overrides
For internal service overrides, aggregate the `_dd.base_service` tag in order to provide stats for the service override broken out by the service from which it was emitted.

This is used to improve visibility of service dependencies for the [Inferred Services beta](https://docs.datadoghq.com/tracing/guide/inferred-service-opt-in/?tab=java).