[apm] fix broken compute_stats_by_span_kind
Fixes a bug introduced in https://github.com/DataDog/datadog-agent/pull/28107 where the `apm_config.compute_stats_by_span_kind` config setting was not being propagated all the way to the new `SpanConcentrator`.

The reason this wasn't caught by the unit tests is that all of the tests were marking the tests spans as `_dd.measured:1` which meant that stats would be calculated regardless of `span.kind`.