Remove redundant error message
Checks failing to load would produce a separate error message with
more details. Not having any valid checks in a config does not have
any additional consequences other than the checks not working, so we
do not need to explicitly call this out. And this also avoids
incorrect error message when the config only contained jmx instances.