fix: get lambda runtime API endpoint dynamically
Use an `ORIGINAL_AWS_LAMBDA_RUNTIME` environment variable to obtain the
real value of `AWS_LAMBDA_RUNTIME`, so the runtime API endpoint does not
need to be hardcoded.

This is necessary with new runtimes that don't necessarily use the usual
`127.0.0.1:9001` endpoint.

The extension wrapper script fowards the original `AWS_LAMBDA_RUNTIME`
to `ORIGINAL_AWS_LAMBDA_RUNTIME` before overwriting it.