usm: Move test libraries to site-packages/ddtrace
We need some shared libraries in specific paths in order to add tests
for the Python APM instrumentation detection in service discovery.
Using symbolic links does not work.  Since the path name does not matter
for this eBPF SSL tests, rename the path to the one needed by the
upcoming service discovery tests.