Merge mocks for the config from comp and pkg into a single one
This prevent issue where the two mocks were created and would hold
different configuration to different part of the code base.

This also align the way we use mocks for configuration across the code
base.

Finally this is the first step toward deprecated the use of Fx for mocks
in the code base. For more see
https://datadoghq.dev/datadog-agent/components/testing/.