discovery: Report language from system-probe
Currently, (cmdline-based) language detection is being done both in
system-probe (since its a requirement for instrumentation detection
which needs to be done there due to filesystem access) and from the
check (which also has access to the cmdline).  Do it only from the
former since it will need elevated privileges when it starts to look at
the binary.