discovery: Fix service name extraction from containers
Language/framework-specific service name generation does not currently
work from containers since the code does not read the file from the
containers' root.

The service name generation code already uses the fs.FS interfaces
(according to the comments, for testing) so these can be used to ensure
that filesystem access happen inside the container's root.  However, a
problem with the fs.FS interface is that it does not allow absolute
paths nor paths with .. or ./ etc to be passed to Open/Stat/etc but the
service name generation code can do this in several places.  To handle
this, we have our own fs.FS interface implementation with wraps fs.DirFS
to handle this transparently for the callers.

In order to ensure that this code is exercised in the tests, the
existing "testdata" directories are moved into "testdata/root", and the
fs.FS mechanism is used to set the latter as the root directory during
testing.

With this in place, make system-probe pass the /proc/$PID/root as the
root to get the service name generation to work for containers.