languagedetection: Read binID directly from exe
The current code tries to first read the target of the exe symbolic link
and then appends that path to the proc's root, and then tries to stat(2)
that.  That doesn't work if the binary is not available inside the
proc's root for any reason.  However, stat(2)ing the exe itself always
works, so do that instead.

For example avahi-daemon works like this:

 $ docker run --privileged -it --rm --platform linux/amd64 -v
 /proc/:/host/proc/:ro  --entrypoint /bin/sh alpine
 / # ls -l /host/proc/1640/exe
 lrwxrwxrwx    1 root     root             0 Aug 28 09:09
 /host/proc/1640/exe -> /usr/sbin/avahi-daemon
 / # ls -l /host/proc/1640/root
 lrwxrwxrwx    1 root     root             0 Aug 28 12:38
 /host/proc/1640/root -> /etc/avahi
 / # ls /host/proc/1640/root/usr/sbin/avahi-daemon
 ls: /host/proc/1640/root/usr/sbin/avahi-daemon: No such file or directory
 / # stat -L /host/proc/1640/exe
   File: /host/proc/1640/exe
   Size: 154120          Blocks: 304        IO Block: 4096   regular file
 Device: fc04h/64516d    Inode: 3932513     Links: 1

On the host:
 $ stat /usr/sbin/avahi-daemon
   File: /usr/sbin/avahi-daemon
   Size: 154120    	Blocks: 304        IO Block: 4096   regular file
 Device: fc04h/64516d	Inode: 3932513     Links: 1

As can be seen, stat(2) on the exe gets the correct details of the
binary while the current method doesn't work.