binspect: Fix EOF handling
If a symbol to be found has its name as the last string in the string
table, and the string table is at the end of the file, and the symbol's
length is less than the maximum length of the symbols we are searching
for, the code fails to find this symbol, since ReadAt() returns EOF and
a size lesser than the size of the preallocated array.

To fix this, ignore EOF and instead use the return value of ReadAt() to
known how much to process.  This works if ReadAt() returns lesser than
the prealloc buffer's size, and also handles the case when ReadAt()
returns 0 (with err = EOF).