discovery: Use service list for aliveness check
Currently, the agent check uses the all process list to check if
the process is alive or not, but we can use the service list
instead, since that also considers open ports.  So if a process
closes its listening port the service is no longer alive.

This also allows to use the extra fields received from system-probe
such as RSS and allows the proc usage in the check to be eliminated
in the future.