regression detector: rig up JUnit tgz for CI tools
In preparation for uploading JUnit XML files to Test Visibility via
Agent's CI tooling, this commit adds the various commands needed to
build up a `junit-*.tgs` file needed for that tooling.

A call to the shell script used to upload the JUnit gzipped tarball is
also added in this commit, but is commented out in order to do some
local testing.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>