regression detector: use single quotes for printf
In the `printf` call in the Regression Detector's GitLab task
definition, I neglected to use single quotes for its string argument,
and used double quotes instead, which caused `printf` to interpret
part of the string as a flag (specifically, the "--tags" part). This
commit replaces those double quotes with single quotes so that the
correct string is printed.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>