regression detector: datadog-ci upload junit.xml
We'd like to upload the `junit.xml` file generated by SMP to Test
Visibility, so this commit adds commands to do so. Note that we do not
use the `invoke -e junit_upload` task because it includes logic
inapplicable to the `junit.xml` files generated by SMP, so the
implementation instead invokes `datadog-ci junit upload` directly.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>