smp: stop timeouts due to stale merge-bases
If the Regression Detector tests a PR with a merge-base more than 4
days old, then the container image for the merge-base commit that SMP
would use as a baseline has expired from SMP's ECR. That container
image's absence would eventually cause a GitLab CI timeout, but timing
out in this situation is not useful -- it wastes time, wastes compute,
and provides no insight to users as to the cause of the timeout error.

This commit improve CI operation by checking whether each commit used
in querying SMP's ECR is recent enough that the image has not expired
from SMP's ECR. If the commit is not recent enough, then the CI step
exits with an error message and an exit code of 1. The error message
is intended to be informative and suggest a way to fix the error, but
may need feedback to improve it.

Technically, the time that should be checked is that the image is no
more than 4 days old at the point SMP queries the image. We cannot
know precisely when that will happen, so we check whether an image is
no more than 4 days minus 1 hour old instead. The "minus 1 hour"
portion is an overestimate of the time it would take for SMP to pull
the image, and guards against a pathological, hard-to-diagnose bug in
which the shell command commit timestamp test passes, yet the image
expires out of ECR before SMP can pull it. This buffer of an hour
could be tweaked if a more precise estimate is needed; for now, the
extra engineering effort required for such an estimate does not seem
worthwhile.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>