regression detector: use pr-commenter service
Use of the `pr-commenter` binary from Datadog's devtools repo seems to
be deprecated in favor of using an internal `pr-commenter`
service. This commit splits the
`single-machine-performance-regression_detector` Agent CI job config
into two jobs: one with the same name, and another called
`single-machine-performance-regression_detector-commenter`.

The first of these two jobs contains all of the commands in the
existing `single-machine-performance-regression_detector` job, except
for installing the `pr-commenter` binary discussed above, and
`pr-commenter`-related commands.

The second of these two jobs is scoped to posting Regression Detector
PR comments using a container that includes `pr-commenter`.

Previous SMP attempts to migrate from the `pr-commenter` binary to a
`pr-commenter` service have been blocked because we were unable to
figure out how to provide the necessary authentication for that
service. I believe that the `pr-commenter` container includes a
mechanism for authentication that should unblock this
migration. However, if testing reveals that this approach fails due to
lack of authentication, an alternative approach may be necessary.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>