regression detector: use docker runner
I don't know enough about the details of the Agent CI pipeline to
understand how to configure a CI job so it uses a Kubernetes
runner. Whatever I did in the previous commit does not seem to trigger
CI, so this commit explicitly tags the
`single-machine-performance-regression_detector-commenter` job with
`"runner:docker"` in addition to `"arch:amd64"`.

Two additional considerations come into play:

First, I discovered that the GitLab CI YAML `needs:` keyword does
*not* download artifacts by default, but the commenter job needs
access to those artifacts, so this commit reworks the `needs:` section
to include `artifacts:true` so that the commenter job downloads the
artifacts from the Regression Detector job.

Second, I suspect the Docker runner uses different syntax to disable
the entrypoint script of a Docker container. The commenter job is
modeled on the `golang_deps_diff` job, which includes an environment
variable that disables the entrypoint script of the Docker image that
job uses. So this commit removes the now seemingly unnecessary
Kubernetes environment variables, and adds an `image:entrypoint`
setting to disable the entrypoint script.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>