regression detector: make transformed report file
It would be easier to debug the Markdown report to valid JSON string
transformation if the output of that transformation were written
somewhere, like to a file. This commit refactors PR comment service
JSON payload creation so the JSON string transformation output is
written to a text file, and that text file is stored in the job's
artifacts. If bugs occur, the file can be downloaded to diagnose bugs.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>