regression detector: remove spaces around colons
For reasons not entirely clear to me, removing the spaces after colons
in the strings I'm using to construct the JSON payloads seems to
mollify whatever YAML parser GitLab (and VSCode?) are using. Could be
a rule I am running afoul of without realizing it -- I am not
conversant in the YAML 1.x specs.

Anyway, this commit removes those spaces, and hopefully that resolves
the server-side YAML error I am seeing.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>